% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgbump.R
\name{pkgbump}
\alias{pkgbump}
\title{Update Package Version Number}
\usage{
pkgbump(version)
}
\arguments{
\item{version}{version number to write}
}
\value{
Update package version number in tracked files
}
\description{
Auto-increment package version number for across tracked package files.
This was originally intented to update version numbers in the DESCRIPTION,
package.json, and use_*.R files. Use the function \code{set_pkgbump} to define
which files should be mananaged by pkgbump.
}
\examples{
\dontrun{
  pkgbump(version = "1.0.0")
}

}
